Allow arbitrary thread numbers per work group in likwid-bench
Up to now, different work groups must have the same thread number for
likwid-bench to work properly. With this change, different work groups
can have different numbers of threads, which is useful for e.g., when
one wants to have 2 threads on the first socket and only one on the
second.