Fix memory corruption in sysFeatures_list
The strings were incorrectly copied when their length exceeded
HWFEATURES_MAX_STR_LENGTH. To solve this we just use strndup, as it is
te perfect fit for correctly copying strings at a maximum size.

In addition adding a feature to an internal feature list used to ignore
the 'tester' and 'unit' field. This would cause corruption later on.