Add likwid_device_create_from_string function
This almost an entire rewrite of the old likwid_device_create function.
The intention is to support PCI addresses and index numbers, but the
former ones cannot be supported with the previous API.

To avoid a lot of duplicated code, the old likwid_device_create function
now shares the important parts of the code with
likwid_device_create_from_string.