Add device string support for likwid-sysfeatures
These device strings don't just allow to specify hardware threads,
but also allow GPU devices to be declared.

This slightly changes the behavior of likwid-sysfeatures, as it no
longer queries the devices in the upper device heirarchy (e.g. core and
node for a hardware thread) when just that device name is specified.

Each hierarchy level now needs to be specified explicitly.