Remove unnecessary parenthesis from dlopen macros
Refactor cuda/rocm error handling

Main background of this change is actually clang-format.
The problem is that we use a not 100% safe formatting rule, which
removes unnecessary parentheses. One of those cases where this rule is
not safe is the following:

_define mymacro(func, args) func args

While this type of macro already looks fishy, it relies on the 'args'
being passed with parantheses. Now, if only a single argument is passed,
those will be removed, breaking the functionality of this macro.