rocmon: postpone HIP initialization after rocprofiler_force_configure
Using HIP before that makes rocprofiler-sdk not initialize properly and
it will complain that HSA is not initialized.

Unforuntately this requires the code to become more ugly, since we have
to filter out the "unwanted" (aka HIP_VISIBLE_DEVICES) GPUs at a later time.
Accordingly we keep track of all devices, but mark them whether they are
used or not, and which HIP ID they have (if applicable).